+++ CHANGELOG.md
@@ -4,3 +4,4 @@ Once we have a solid base, we will bump to v1 without any breaking changes from 
 
 * #13 Fixed export writing task confusing success and failure (Thanks, @filmor)
 * #10 Fixed a bug where loading a NXPorts enabled project in designtime build mode (e.g. Visual Studio) would be blocked by a failing AssemblyExportWriterTask
+* #14 NXPorts now removes the obsolete `NXPorts.Attributes.dll` from the build output